\name{knit-package}
\alias{knit-package}
\alias{knit}
\docType{package}
\title{
\packageTitle{knit}
}
\description{
\packageDescription{knit}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{knit}
\packageIndices{knit}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{knit}

Maintainer: \packageMaintainer{knit}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
